segmentation/requirements.txt
torch==2.0.1
torchvision==0.15.2
h5py>=3.6.0
scipy>=1.7.3
opencv-python>=4.5.5
timm>=0.5.4
albumentations>=1.1.0
tensorboardX>=2.4.1
gdown>=4.2.1
mmsegmentation==0.30.0
#mmcv==1.6.2
